requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai
python-dotenv==1.0.0
